Components: rework EffectUnit UX
Replace showParametersButton with focusEffectButton. This button's LED now indicates whether
an effect is focused.
* Short press, parameters hidden: show parameters
* Short press, parameters showing: unfocus and hide parameters. The focused effect is remembered and restored
  the next time parameters are shown.
* Long press, parameters hidden or showing: switch to effect focus choose mode

In effect focus choose mode, the effect enable buttons act to select the focused effect. Their LEDs
show whether that effect is focused. When focusEffectButton is released, the effect buttons go back
to controlling the enable switches and their LEDs are updated to reflect which effects are enabled.