use QVector<uint8_t> instead of QByteArray for controller input
Converting QByteArray to a QJSValue turns it into an ArrayBuffer in JS,
which then needs to be converted to a Uint8Array to be useful.
QVector<uint8_t> is converted directly to a JS Array.

This temporarily disables debugging for MIDI System Exclusive messages.
Fixing this will require refactoring sending controller output to also
use QVector<uint8_t> instead of QByteArray, which would make this commit
huge.